requirements.txt
asgiref==3.9.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==2.1.0
dj-rest-auth==7.0.1
distlib==0.4.0
Django==5.2.5
django-allauth==65.13.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.20.0
gunicorn==25.3.0
idna==3.11
packaging==26.2
pillow==12.2.0
platformdirs==4.5.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.2
virtualenv==20.35.3
whitenoise==6.12.0
